refactor file structure and tests
