+++ lib/grid.js
@@ -1,6 +1,35 @@
 var $ = require('jquery');
-var GameGrid = require('../lib/tile.js');
 var Tile = require('../lib/tile.js');
+
+
+function linkedTiles(){
+  this._length = 0;
+  this.head = null;
+  this.tail = null;
+}
+
+var allTiles = [];
+
+for (var i = 0; i < 16; i++) {
+  allTiles.push(new Tile('2', i));
+}
+
+allTiles.forEach(function(e, i) {
+  e.next = allTiles[i + 1];
+  if (i > 0) {
+    e.previous = allTiles[i - 1];
+  }
+  if (i > 3) {
+    e.above = allTiles[i-4];
+  }
+  if (i < 12) {
+    e.below = allTiles[i+4];
+  }
+});
+
+var head = allTiles[0];
+
+console.log(head);
 
 // create pullTiles
 // when the up arrow is pressed -- key 38
@@ -38,7 +67,7 @@ var Tile = require('../lib/tile.js');
 var pullTile;
 
 function smush() {
-  pullTile.data = pullTile.next.data;
+  pullTile.data =   pullTile.next.data;
   pullTile.next.data = null;
 }
 
@@ -47,4 +76,5 @@ function multiply() {
 }
 
 
-module.exports = GameGrid;
+
+module.exports = allTiles;
