+++ lib/tile.js
@@ -1,5 +1,4 @@
 var $ = require('jquery');
-
 // var _ = require('lodash');
 
 // Since every instance of a DoublyList is instantiated without nodes, the default values of head and tail are set to null.
@@ -12,15 +11,20 @@ function Tile (data, position) {
   this.below = null;
 }
 
-function linkedTiles(){
-  this._length = 0;
-  this.head = null;
-  this.tail = null;
-}
-
 Tile.prototype.start = function() {
   return this.data = 2;
 };
+
+var tile1 = new Tile(null, 1);
+$('#c1').text(tile1.data);
+tile1.next = new Tile(null, 2);
+$('#c2').text(tile1.next.data);
+tile1.next.next = new Tile(null,3);
+$('#c3').text(tile1.next.next.data);
+tile1.next.next.next = new Tile(null,4);
+$('#c4').text(tile1.next.next.next.data);
+
+tile1.start();
 
 
 Tile.prototype.moveRight = function(tile1, tile2) {
@@ -33,8 +37,7 @@ Tile.prototype.moveRight = function(tile1, tile2) {
 };
 
 
-// what that file exports, otherwise everything in this file is private
-// module.exports = makeFunctions;
+
 // you can export functions or objects
 // you only want one module.exports per file
 //
